requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
